+++ packages/client/src/assets/CheckIcon.tsx
@@ -1,0 +1,8 @@
+export const CheckIcon = () => (
+  <svg width="35" height="35" fill="currentColor" viewBox="0 0 24 24">
+    <path
+      fill="currentColor"
+      d="M9 18.25a.74.74 0 0 1-.53-.25l-5-5a.75.75 0 1 1 1.06-1L9 16.44 19.47 6a.75.75 0 0 1 1.06 1l-11 11a.74.74 0 0 1-.53.25Z"
+    />
+  </svg>
+)
