+++ packages/client/src/assets/SoundOffIcon.tsx
@@ -1,0 +1,14 @@
+export const SoundOffIcon = () => (
+  <svg width="35" height="35" viewBox="0 0 50 50">
+    <path
+      d="m35.735 24.4 4.542-4.548a1.001 1.001 0 0 0-1.416-1.414l-4.545 
+					4.552-4.545-4.552a1.001 1.001 0 0 0-1.416 1.414l4.542 4.548-4.542 
+					4.547a1.001 1.001 0 0 0 1.416 1.414l4.545-4.551 4.545 4.551a1 1 0 
+					0 0 1.416-1.414L35.735 24.4zM22.642 38.06a1.002 1.002 0 0 0 
+					1-1v-24c0-.379-.214-.725-.552-.895a1.005 1.005 0 0 0-1.047.094l-8.2 
+					6.134H9.569a1 1 0 0 0-1 1v11.334a1 1 0 0 0 1 1h4.273l8.2 
+					6.134c.177.131.388.199.6.199zm-1-2.997-6.867-5.137a1.002 1.002 0 0 
+					0-.599-.199H10.57v-9.334h3.606c.216 0 .426-.07.599-.199l6.867-5.137v20.006z"
+    />
+  </svg>
+)
