--- packages/client/src/entities/before-game/assets.ts
@@ -1,14 +1,0 @@
-export const colors = [
-  { color: '#b38ff1' }, //--color-background-light
-  { color: '#6495ED' }, //CornflowerBlue
-  { color: '#006400' }, // зеленый
-  { color: '#d61e1e' }, // --color-contrast
-  { color: '#778899' }, //LightSlateGray
-  { color: '#F08080' }, //LightCoral
-]
-
-export const cards = [
-  { image: 'sprites/card.png' },
-  { image: 'sprites/card2.png' },
-  { image: 'sprites/card3.png' },
-]
