+++ packages/client/src/features/game/assets.ts
@@ -1,5 +1,4 @@
 export const LOGO_HEIGHT = 95
-export const closedCardImage = 'sprites/card.png'
 
 export const suits = {
   clubs: 'clubs',
@@ -20,6 +19,21 @@ export const button_text = {
   HeTake: 'Пусть берет',
   Ok: 'Бито',
 }
+
+export const colors = [
+  { color: '#b38ff1' }, //--color-background-light
+  { color: '#6495ED' }, //CornflowerBlue
+  { color: '#006400' }, // зеленый
+  { color: '#d61e1e' }, // --color-contrast
+  { color: '#778899' }, //LightSlateGray
+  { color: '#F08080' }, //LightCoral
+]
+
+export const cards = [
+  { image: 'sprites/card.png' },
+  { image: 'sprites/card2.png' },
+  { image: 'sprites/card3.png' },
+]
 
 export const initialDeckCard = [
   {
