+++ packages/client/src/features/game/styles.module.css
@@ -1,30 +1,36 @@
-.container {
+.game,
+.gameBoard {
   display: flex;
   justify-content: center;
-  width: 100vw;
-  height: 100vh;
+  align-items: center;
+  box-sizing: border-box;
+}
+
+.game {
+  height: 100%;
+  margin: 95px 5% 3% 5%;
 }
 
 .gameBoard {
-  display: flex;
-  box-sizing: border-box;
-  justify-content: center;
-  align-items: center;
   position: relative;
-  margin: 95px 5% 3% 5%;
   width: 100%;
+  height: 100%;
   border-radius: 10px;
   z-index: 1;
   border: 1px solid var(--color-secondary);
 }
 
-canvas {
+canvas,
+.button,
+.noticeToPlayer {
   position: absolute;
+}
+
+canvas {
   top: 0;
 }
 
 .button {
-  position: absolute;
   top: 70%;
   right: 20%;
 }
@@ -38,12 +44,6 @@ canvas {
 }
 
 .noticeToPlayer {
-  position: absolute;
   bottom: 20%;
   left: 31%;
-}
-
-.noticeToBot {
-  position: absolute;
-  top: 20%;
 }
