+++ packages/client/src/pages/game-page/GamePage.tsx
@@ -1,5 +1,12 @@
 import { Game } from '../../features/game/Game'
+import { ButtonGoBack } from '../../shared/button-go-back'
+import styles from './styles.module.css'
 
 export const GamePage = () => {
-  return <Game />
+  return (
+    <div className={styles.gamePage}>
+      <ButtonGoBack />
+      <Game />
+    </div>
+  )
 }
