+++ packages/client/src/pages/game-page/styles.module.css
@@ -1,0 +1,6 @@
+.gamePage {
+  display: flex;
+  flex-direction: column;
+  width: 100vw;
+  height: 100vh;
+}
